[S13P21A302-96] 소셜 로그인 구현

See merge request s13-bigdata-dist-sub1/S13P21A302!11

+++ backend-spring/leaper/src/main/java/com/ssafy/leaper/domain/insight/dto/response/dailyAccountInsight/DailyAccountViewsResponse.java
@@ -5,10 +5,10 @@ import java.math.BigInteger;
 import java.time.LocalDate;
 
 public record DailyAccountViewsResponse(
-    Long platformAccountId,
-    String platformTypeId,
-    BigInteger totalViews,
-    LocalDate snapshotDate
+        Integer platformAccountId,
+        String platformTypeId,
+        BigInteger totalViews,
+        LocalDate snapshotDate
 ) {
   public static DailyAccountViewsResponse from(DailyAccountInsight dailyAccountInsight){
     return new DailyAccountViewsResponse(
